node/PROJECT/NEST-CLEAN/src/domain/forum/application/use-cases/register-student.ts
import { Either, left, right } from "@/core/either";
import { Injectable } from "@nestjs/common";
import { Student } from "../../enterprise/entities/student";
import { HashGenerator } from "../cryptography/hash-generator";
import { StudentsRepository } from "../repositories/student-repository";
import { StudentAlreadyExistsError } from "./errors/student-already-exists-error";

interface RegisterStudentUseCaseRequest {
  name: string;
  email: string;
  password: string;
}

type RegisterStudentUseCaseResponse = Either<
  StudentAlreadyExistsError,
  {
    student: Student;
  }
>;

@Injectable()
export class RegisterStudentUseCase {
  constructor(
    private studentsRepository: StudentsRepository,
    private hashGenerator: HashGenerator
  ) {}

  async execute({
    name,
    email,
    password,
  }: RegisterStudentUseCaseRequest): Promise<RegisterStudentUseCaseResponse> {
    const studentWithSameEmail =
      await this.studentsRepository.findByEmail(email);

    if (studentWithSameEmail) {
      throw left(new StudentAlreadyExistsError(email));
    }

    const password_hash = await this.hashGenerator.hash(password);

    const student = Student.create({
      email,
      name,
      password: password_hash,
    });

    await this.studentsRepository.create(student);

    return right({ student });
  }
}
